create new module for platform implementations, and move ValueClassSupport into it (deleting the duplicated versions in Android and DSL modules)
