+++ modules/mockk-agent-android/build.gradle.kts
@@ -45,6 +45,8 @@ dependencies {
     api(projects.modules.mockkAgentApi)
     api(projects.modules.mockkAgent)
 
+    implementation(projects.modules.mockkPlatformTools)
+
     implementation(kotlin("reflect"))
     implementation("com.linkedin.dexmaker:dexmaker:${buildsrc.config.Deps.Versions.dexmaker}")
     implementation("org.objenesis:objenesis:${buildsrc.config.Deps.Versions.objenesis}")
