+++ settings.gradle.kts
@@ -14,6 +14,7 @@ include(
     ":modules:mockk",
     ":modules:mockk-agent-api",
     ":modules:mockk-agent",
+    ":modules:mockk-platform-tools",
     ":modules:mockk-dsl",
 
     ":test-modules:client-tests",
